+++ app/Http/Controllers/DepartmentController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Domain\Academic\Models\Department;
 use App\Domain\Settings\Models\UnitCategory;
+use App\Domain\Academic\Models\Module;
+use App\Domain\Academic\Models\Program;
 use App\Domain\Settings\Models\Campus;
 use App\Domain\Academic\Actions\DepartmentAction;
 use App\Utils\Util;
@@ -84,8 +86,12 @@ class DepartmentController extends Controller
     {
         try{
             $department = Department::findOrFail($id);
-            $department->delete();
-            return redirect()->back()->with('message','Department deleted successfully');
+            if(Module::where('department_id',$department->id)->count() != 0 || Program::where('department_id',$department->id)->count() != 0){
+               return redirect()->back()->with('message','Department cannot be deleted because it has modules or programs');
+            }else{
+               $department->delete();
+               return redirect()->back()->with('message','Department deleted successfully');
+            }
         }catch(Exception $e){
             return redirect()->back()->with('error','Unable to get the resource specified in this request');
         }
